+++ src/es6/index.js
@@ -1,3 +1,3 @@
 
-export { Series, DataFrame } from './core/index';
+export { Series, DataFrame, concat } from './core/index';
 export { to_datetime } from './tseries/tools';